refactor: move request code from LiveServiceDemo to request-service lib
